fix(example): remove unused code + resources; add branding
