+++ build.gradle
@@ -1,17 +1,22 @@
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
-
 buildscript {
-    ext.kotlin_version = '1.2.61'
+    ext {
+        kotlinVersion = '1.2.61'
+        buildToolsVersion = '28.0.2'
+        minSdkVersion = 16
+        targetSdkVersion = 28
+        compileSdkVersion = 28
+
+        versionName = '0.0.0'
+        versionCode = 0
+    }
+
     repositories {
         google()
         jcenter()
     }
     dependencies {
         classpath 'com.android.tools.build:gradle:3.2.0-rc03'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-
-        // NOTE: Do not place your application dependencies here; they belong
-        // in the individual module build.gradle files
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlinVersion"
     }
 }
