+++ example/src/main/res/values/drawables.xml
