+++ example/src/main/res/values/strings.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+
+    <string name="app_name">TreeRecyclerView Example</string>
+    <string name="main_drawer_open">Open navigation drawer</string>
+    <string name="main_drawer_close">Close navigation drawer</string>
+
+    <string name="drawer_info_creator">Created by Jonas Wanke</string>
+    <string name="drawer_info_contact">contact@jonas-wanke.com</string>
+
+</resources>
+
